backend/requirements.txt
pillow
rembg
openai
google-genai
fastapi
uvicorn[standard]
